Fix removed tag actions not being logged in the undo/redo log
When a tag was removed this action was not logged in the undo/redo log and as such undo-ing a remove-tag action was impossible. 

The reason for this bug was a missing `await`.

This change fixes this by adding the missing `await` which causes the undo action to be added to the undo/redo log.